project linted

+++ contracts/IGasPrice.sol
@@ -3,5 +3,5 @@
 pragma solidity ^0.6.12;
 
 interface IGasPrice {
-    function maxGasPrice() external returns (uint);
+    function maxGasPrice() external returns (uint256);
 }
